Modify download logic for JSON configuration files
Add conditional logic to download different JSON files based on test mode.

+++ functions_t.sh
@@ -5300,7 +5300,12 @@ function my() {
       if [ "$MACHINE" = "VIRTUAL" ] && [ "$HYPERVISOR" = "KVM" ]; then
         curl -skL# https://raw.githubusercontent.com/PeterSuh-Q3/redpill-load/master/config/pats_t.json -o $configfile
       else  
-        curl -skL# https://raw.githubusercontent.com/PeterSuh-Q3/redpill-load/master/config/pats.json -o $configfile
+        if [ -f /tmp/test_mode ]; then
+            cecho g "###############################  This is Test Mode  ############################"        
+            curl -skL# https://raw.githubusercontent.com/PeterSuh-Q3/redpill-load/master/config/pats_t.json -o $configfile
+        else
+            curl -skL# https://raw.githubusercontent.com/PeterSuh-Q3/redpill-load/master/config/pats.json -o $configfile
+        fi    
       fi  
       echo "offline = ${offline}"
         [ "${offline}" = "NO" ] && _pat_process
